requirements.txt
youtube-transcript-api>=0.5.0
google-adk>=0.1.0
python-dotenv>=1.0.0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
